Automatic mocking of inherited traits
Previously proper mocking was possible using multi-trait form of
`mock!` macro only. You had to provide definitions for mocked trait
as well as all base traits.

Now it is possible to just use `#[derive_mock]` on all base traits
and target one and it will work. However, it only works if all traits
are defined locally and used without global paths. This is yet to
be fixed.